golang.bbclass: filter out vendored packages from build
While it would be nice if the go tool did this automatically
for us when given a wildcard on the command line, it doesn't
and never will.  So instead, add a function that will do
that filtering for us.

This will let us place the vendor directory directly at
${GO_SRCROOT} rather than above it, so vendored sources
can be included with the rest of the sources in the
do_install step.  For #7.

Signed-off-by: Matt Madison <matt@madison.systems>